refactor: request 패키지로 이동 (#16)
