+++ src/main/java/com/chungbazi/server/domain/user/api/dto/request/UserNameRequest.java
@@ -1,4 +1,4 @@
-package com.chungbazi.server.domain.user.api.dto;
+package com.chungbazi.server.domain.user.api.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
