+++ src/main/java/com/chungbazi/server/domain/user/application/validator/UserValidator.java
@@ -4,8 +4,8 @@ import com.chungbazi.server.domain.policy.domain.type.EducationCode;
 import com.chungbazi.server.domain.policy.domain.type.EmploymentCode;
 import com.chungbazi.server.domain.policy.domain.type.PolicySubCategoryType;
 import com.chungbazi.server.domain.policy.domain.type.SidoCode;
-import com.chungbazi.server.domain.user.api.dto.UserOnboardingRequest;
-import com.chungbazi.server.domain.user.api.dto.UserPolicyRequest;
+import com.chungbazi.server.domain.user.api.dto.request.UserOnboardingRequest;
+import com.chungbazi.server.domain.user.api.dto.request.UserPolicyRequest;
 import com.chungbazi.server.domain.user.domain.type.IncomeLevel;
 import com.chungbazi.server.domain.user.exception.UserException;
 import com.chungbazi.server.domain.user.exception.code.UserErrorCode;
